Turn HealthSavingsAccount into a generic interface.
Different parts of the code need to use the start/end date of HSAs
as either a Date object or a string. Due to a limitation of the
schema generator used to convert typescript interfaces into JSON
schemas, a Date type is coverted into a string type in the schema
with a 'date-time' formatter. This means that the central information
interface must have these dates specified as a string, however
when actaully working with HealthSavingsAccount objects we want to
convert them back into Date objects.